Move session iframe pages
Closes #14767
