+++ services/src/main/resources/org/keycloak/protocol/oidc/endpoints/3p-cookies-step2.html
